+++ static-site-generator/package.json
@@ -30,19 +30,15 @@
     "last 1 ChromeAndroid versions"
   ],
   "devDependencies": {
-    "@storybook/addon-actions": "^7.5.3",
-    "@storybook/addon-essentials": "^7.5.3",
-    "@storybook/addon-interactions": "^7.5.3",
-    "@storybook/addon-links": "^7.5.3",
-    "@storybook/react": "^7.5.3",
-    "@storybook/react-vite": "^7.5.3",
-    "@storybook/testing-library": "^0.2.2",
+    "@storybook/addon-docs": "^10.2.7",
+    "@storybook/addon-links": "^10.2.7",
+    "@storybook/react-vite": "^10.2.7",
     "@types/node": "^20.10.0",
     "@types/react": "^18.2.38",
     "@types/react-dom": "^18.2.16",
     "@typescript-eslint/eslint-plugin": "^6.12.0",
     "@typescript-eslint/parser": "^6.12.0",
-    "@vitejs/plugin-react": "^4.2.0",
+    "@vitejs/plugin-react": "^4.7.0",
     "classnames": "^2.3.1",
     "eslint": "^8.54.0",
     "eslint-config-prettier": "^9.0.0",
@@ -56,7 +52,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sass": "^1.69.5",
-    "storybook": "^7.5.3",
+    "storybook": "^10.2.7",
     "stylelint": "^15.11.0",
     "stylelint-config-standard-scss": "^11.1.0",
     "stylelint-order": "^6.0.3",
