+++ static-site-generator/src/stories/components/layout/Inner/index.stories.tsx
@@ -1,23 +1,23 @@
-import type { Meta, StoryObj } from '@storybook/react'
-import { LayoutSection } from '.'
+import type { Meta, StoryObj } from '@storybook/react-vite'
+import { LayoutInner } from '~/components/layout/Inner'
 
 export default {
-  title: 'components/layout/Section',
-  component: LayoutSection,
-} as Meta<typeof LayoutSection>
+  title: 'components/layout/Inner',
+  component: LayoutInner,
+} as Meta<typeof LayoutInner>
 
-type Story = StoryObj<typeof LayoutSection>
+type Story = StoryObj<typeof LayoutInner>
 
 export const Default: Story = {
   render: (args) => {
     return (
-      <LayoutSection {...args}>
+      <LayoutInner {...args}>
         <>
           コンテンツが入ります
           <br />
           テキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキストテキスト
         </>
-      </LayoutSection>
+      </LayoutInner>
     )
   },
 }
